Removed the auto-rejoin code (its in irc_net.c now)
Made a special case for the 437 numeric, because it can be either nickname
or channel related.

Added code so that while a client is detached, we attempt to rejoin channels
if its full/invite only/banned/juped

Also on join, we can reactive a channel, and then spool the log back to the
client (in case of kick).


git-svn-id: http://dircproxy.googlecode.com/svn/trunk@55 36e7e03e-8437-11dd-8f23-d9c8888c2974